refactor: Realigned the project structure to the recommended
