+++ packages/cli/src/index.ts
@@ -1,9 +1,9 @@
-import type { CodemodOptions } from '../types/index.js';
 import {
   createFilesFromBlueprints,
   createOptions,
   updatePackageJson,
 } from './steps/index.js';
+import type { CodemodOptions } from './types/index.js';
 
 export function createCodemod(codemodOptions: CodemodOptions): void {
   const options = createOptions(codemodOptions);
